Adding check in case of a cp failure

+++ src/init/wazuh/wazuh.sh
@@ -102,7 +102,11 @@ WazuhUpgrade()
         cp $DIRECTORY/var/db/global.db* $DIRECTORY/queue/db/
         chmod 640 $DIRECTORY/queue/db/global.db*
         chown ossec:ossec $DIRECTORY/queue/db/global.db*
-        rm -f $DIRECTORY/var/db/global.db*
+        if [ -f "$DIRECTORY/queue/db/global.db" ]; then
+            rm -f $DIRECTORY/var/db/global.db*
+        else
+            echo "Failed to move global.db during the upgrade"
+        fi
     fi
     rm -f $DIRECTORY/var/db/.profile.db*
     rm -f $DIRECTORY/var/db/.template.db*
